+++ src/pages/expenses.retrospective.index.tsx
@@ -1,0 +1,29 @@
+import { createFileRoute } from '@tanstack/react-router';
+import React from 'react';
+
+import AuthGuard from '@/features/auth/ui/AuthGuard';
+import UnReviewedExpenseView from '@/features/expense/ui/UnReviewedExpenseView';
+import RetrospectiveView from '@/features/retrospective/ui/RetrospectiveView';
+import Logo from '@/shared/ui/icons/Logo';
+import BottomNavBar from '@/widgets/BottomNavBar';
+import ReviewTopNavBar from '@/widgets/ReviewTopNavBar';
+
+export const Route = createFileRoute('/expenses/retrospective/')({
+  component: () => (
+    <AuthGuard>
+      <RouteComponent />
+    </AuthGuard>
+  ),
+});
+
+function RouteComponent() {
+  return (
+    <>
+      <header className='px-5 py-4'>
+        <Logo />
+      </header>
+      <RetrospectiveView />
+      <BottomNavBar variant='white' />
+    </>
+  );
+}
